Update readme.md
It might not be bad to keep some sort of summarization of the benefits of this approach as a bullet (serving the same role as the "These three blocks can be combined to process long lists quickly and creatively in various programming scenarios" point). I removed the original one because I am wary about claiming performance unilaterally since this depends on other things like the performance of the runtime, framework (like spark), size of the data, etc.

Personally, I like using functional programming concepts like these because, when applicable, they can often convey the intention of the programmer more easily and can simplify the code. Simpler code is generally has less bugs ;)